Move v2 and version tests to routes folder to match app structure
